Crate massa_serialization

Macros

gen_varint 🔒

Structs

BoolDeserializer

Deserializer for bool

BoolSerializer

Serializer for bool

DeserializeError

OptionDeserializer

OptionSerializer

RatioDeserializer

RatioSerializer

Serializer for Ratio

U16VarIntDeserializer

Deserializer for u16 in a varint form.

U16VarIntSerializer

Serializer for u16 in a varint form.

U32VarIntDeserializer

Deserializer for u32 in a varint form.

U32VarIntSerializer

Serializer for u32 in a varint form.

U64VarIntDeserializer

Deserializer for u64 in a varint form.

U64VarIntSerializer

Serializer for u64 in a varint form.

U128VarIntDeserializer

Deserializer for u128 in a varint form.

U128VarIntSerializer

Serializer for u128 in a varint form.

Enums

SerializeError

Traits

Deserializer

Trait that define the deserialize method that must be implemented for all types have serialize form in Massa.

Serializer

This trait must be implemented to serializes all data in Massa.

Functions

buf_to_array_ctr